complate petrol pump

+++ app/Http/Controllers/admin/ProductController.php
@@ -8,6 +8,7 @@ use App\Product;
 use App\ProductItem;
 use App\CompanyInfo;
 use App\ProductStock;
+use App\Calculation;
 use Auth;
 
 class ProductController extends Controller{
@@ -41,7 +42,7 @@ class ProductController extends Controller{
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request){
-        $validatedData = $request->validate([
+          $validatedData = $request->validate([
             'product_item_id'=>'',
             'company_id'=>'required|max:255',
             'user_id'=>'',
@@ -50,6 +51,7 @@ class ProductController extends Controller{
             'oil_stack'=>'required|max:255',
             'oil_price'=>'required',
             'oil_total_price'=>'required',
+            'payment_option'=>'required',
             'stack_date'=>'required|max:255',
             'oil_description'=>'required',
             'status'=>'',
@@ -61,29 +63,75 @@ class ProductController extends Controller{
               $validatedData['status'] = 0;
         }
 
-      $model = new Product();
-      $success = $model->create($validatedData);
-      if ($success) {
-        $oil_stack = ProductStock::where('product_item_id','=',$request->product_item_id)->first();
-        if($oil_stack){
-            $total_stock = $oil_stack->oil_stack+$request->oil_stack;
-            $model = ProductStock::where('product_item_id','=',$request->product_item_id)->first();
-            $model->oil_stack = $total_stock;
-            $model->save();
-        }else{
-             $validatedData = $request->validate([
-                'product_item_id'=>'',
-                'oil_stack'=>'required|max:255',
-                'stack_date'=>'required|max:255'
-            ]);
+      
+        if($request->payment_option == 'investment'){
+            $data = Calculation::where('invest_id','=',1)->first();
+            $total =  $data->total_invest;
+            if($request->oil_total_price > $total){
+                 return response()->json(['success' => false, 'status' => 'danger', 'message' => _lang('
+                Product price must be less than  investment')]);
+            }else{
+               $expense =  $total - $request->oil_total_price;
+               $data->total_invest = $expense;
+               $data->save();
+               $model = new Product();
+                $success = $model->create($validatedData);
+                if ($success) {
+                    $oil_stack = ProductStock::where('product_item_id','=',$request->product_item_id)->first();
+                    if($oil_stack){
+                        $total_stock = $oil_stack->oil_stack+$request->oil_stack;
+                        $model = ProductStock::where('product_item_id','=',$request->product_item_id)->first();
+                        $model->oil_stack = $total_stock;
+                        $model->save();
+                    }else{
+                        $validatedData = $request->validate([
+                            'product_item_id'=>'',
+                            'oil_stack'=>'required|max:255',
+                            'stack_date'=>'required|max:255'
+                        ]);
 
-            $model = new ProductStock();
-            $model->create($validatedData);
+                        $model = new ProductStock();
+                        $model->create($validatedData);
+                    }
+                    
+                }
+            }
+        }elseif ($request->payment_option == 'savings') {
+            $model = Calculation::where('income_id','=',1)->first();
+            $total =  $model->total_income;
+            if($request->oil_total_price > $total){
+                 return response()->json(['success' => false, 'status' => 'danger', 'message' => _lang('
+                Product price must be less than savings')]);
+            }else{
+               $expense =  $total - $request->oil_total_price;
+               $model->total_invest = $expense;
+               $model->save();
+               $model = new Product();
+                $success = $model->create($validatedData);
+                if ($success) {
+                    $oil_stack = ProductStock::where('product_item_id','=',$request->product_item_id)->first();
+                    if($oil_stack){
+                        $total_stock = $oil_stack->oil_stack+$request->oil_stack;
+                        $model = ProductStock::where('product_item_id','=',$request->product_item_id)->first();
+                        $model->oil_stack = $total_stock;
+                        $model->save();
+                    }else{
+                        $validatedData = $request->validate([
+                            'product_item_id'=>'',
+                            'oil_stack'=>'required|max:255',
+                            'stack_date'=>'required|max:255'
+                        ]);
+
+                        $model = new ProductStock();
+                        $model->create($validatedData);
+                    }
+                    
+                }
+            }
         }
-          
-      }
-      return response()->json(['success' => true, 'status' => 'success', 'message' => _lang('
+        return response()->json(['success' => true, 'status' => 'success', 'message' => _lang('
       Product Added Successfuly'), 'goto' => route('admin.product.index')]);
+
     }
 
     /**
@@ -119,44 +167,106 @@ class ProductController extends Controller{
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, $id){
-
+         
             $models = Product::findOrFail($id);
-            $total_stock = $models->oil_stack;
-            $product_stock = ProductStock::where('product_item_id','=', $request->product_item_id)->first();
-            $stock_total = $product_stock->oil_stack - $total_stock;
-            $product_stock->oil_stack =  $stock_total;
-            $product_add = $product_stock->save();
-            if($product_add){
-                $stock_product = ProductStock::where('product_item_id','=', $request->product_item_id)->first();
-                $add_product = $stock_product->oil_stack + $request->oil_stack;
-                if($add_product){
-                    $stock_product->oil_stack = $add_product;
-                    $data=$stock_product->save();
-                    if($data){
-                        $validatedData = $request->validate([
-                            'product_item_id'=>'',
-                            'company_id'=>'required|max:255',
-                            'user_id'=>'',
-                            'vehicle_name'=>'required|max:255',
-                            'vehicle_number'=>'required|max:255',
-                            'oil_stack'=>'required|max:255',
-                            'oil_price'=>'required',
-                            'oil_total_price'=>'required',
-                            'stack_date'=>'required|max:255',
-                            'oil_description'=>'required',
-                            'status'=>'',
-                        ]);
+            if ($request->payment_option=='investment') {
+               $data =  Calculation::where('invest_id','=',1)->first();
+               $invest = $data->total_invest + $models->oil_total_price;
+               $data->total_invest = $invest;
+               $success = $data->save();
+               if($success){
+                   $minus = $data->total_invest - $request->oil_total_price;
+                   $data->total_invest = $minus;
+                  $message =  $data->save();
+                  if($message){
+                    $total_stock = $models->oil_stack;
+                    $product_stock = ProductStock::where('product_item_id','=', $request->product_item_id)->first();
+                    $stock_total = $product_stock->oil_stack - $total_stock;
+                    $product_stock->oil_stack =  $stock_total;
+                    $product_add = $product_stock->save();
+                    if($product_add){
+                        $stock_product = ProductStock::where('product_item_id','=', $request->product_item_id)->first();
+                        $add_product = $stock_product->oil_stack + $request->oil_stack;
+                        if($add_product){
+                            $stock_product->oil_stack = $add_product;
+                            $data=$stock_product->save();
+                            if($data){
+                                $validatedData = $request->validate([
+                                    'product_item_id'=>'',
+                                    'company_id'=>'required|max:255',
+                                    'user_id'=>'',
+                                    'vehicle_name'=>'required|max:255',
+                                    'vehicle_number'=>'required|max:255',
+                                    'oil_stack'=>'required|max:255',
+                                    'oil_price'=>'required',
+                                    'oil_total_price'=>'required',
+                                    'stack_date'=>'required|max:255',
+                                    'oil_description'=>'required',
+                                    'status'=>'',
+                                ]);
 
-                        if ($request->status) {
-                            $validatedData['status'] = 1;
-                        }else{
-                            $validatedData['status'] = 0;
+                                if ($request->status) {
+                                    $validatedData['status'] = 1;
+                                }else{
+                                    $validatedData['status'] = 0;
+                                }
+                                $model =Product::findOrFail($id);
+                                $model->update($validatedData);
+                            }
                         }
-                        $model =Product::findOrFail($id);
-                        $model->update($validatedData);
+                        
                     }
-                }
-                
+                  }
+               }
+            }elseif($request->payment_option=='savings'){
+                $data =  Calculation::where('invest_id','=',1)->first();
+               $invest = $data->total_income + $models->oil_total_price;
+               $data->total_income = $invest;
+               $success = $data->save();
+               if($success){
+                   $minus = $data->total_income - $request->oil_total_price;
+                    $data->total_income = $minus;
+                  $message =  $data->save();
+                  if($message){
+                    $total_stock = $models->oil_stack;
+                    $product_stock = ProductStock::where('product_item_id','=', $request->product_item_id)->first();
+                    $stock_total = $product_stock->oil_stack - $total_stock;
+                    $product_stock->oil_stack =  $stock_total;
+                    $product_add = $product_stock->save();
+                    if($product_add){
+                        $stock_product = ProductStock::where('product_item_id','=', $request->product_item_id)->first();
+                        $add_product = $stock_product->oil_stack + $request->oil_stack;
+                        if($add_product){
+                            $stock_product->oil_stack = $add_product;
+                            $data=$stock_product->save();
+                            if($data){
+                                $validatedData = $request->validate([
+                                    'product_item_id'=>'',
+                                    'company_id'=>'required|max:255',
+                                    'user_id'=>'',
+                                    'vehicle_name'=>'required|max:255',
+                                    'vehicle_number'=>'required|max:255',
+                                    'oil_stack'=>'required|max:255',
+                                    'oil_price'=>'required',
+                                    'oil_total_price'=>'required',
+                                    'stack_date'=>'required|max:255',
+                                    'oil_description'=>'required',
+                                    'status'=>'',
+                                ]);
+
+                                if ($request->status) {
+                                    $validatedData['status'] = 1;
+                                }else{
+                                    $validatedData['status'] = 0;
+                                }
+                                $model =Product::findOrFail($id);
+                                $model->update($validatedData);
+                            }
+                        }
+                        
+                    }
+                  }
+               }
             }
       return response()->json(['success' => true, 'status' => 'success', 'message' => _lang('
       Product Update Successfuly'), 'goto' => route('admin.product.index')]);
@@ -172,14 +282,48 @@ class ProductController extends Controller{
         //dd($request->slug);
         $minus_stock = ProductStock::where('product_item_id','=',$request->slug)->first();
         $model = Product::findOrFail($id);
-        $total_stock = $minus_stock->oil_stack - $model->oil_stack;
-        $minus_stock->oil_stack = $total_stock;
-        $result = $minus_stock->save();
-        if($result){
-            $models = Product::findOrFail($id);
-            $models->delete();
-            return response()->json(['success' => true, 'status' => 'success', 'message' => _lang('Product Delete  Successfuly'), 'goto' => route('admin.product.index')]);
+        if($model->payment_option == 'investment'){
+            $invest = Calculation::where('invest_id','=',1)->first();
+            $total = $invest->total_invest +  $model->oil_total_price;
+            $invest->total_invest = $total;
+            $success = $invest->save();
+           if($success){
+               $total_stock = $minus_stock->oil_stack - $model->oil_stack;
+                $minus_stock->oil_stack = $total_stock;
+                $result = $minus_stock->save();
+                if($result){
+                    $models = Product::findOrFail($id);
+                    $models->delete();
+                    return response()->json(['success' => false, 'status' => 'danger', 'message' => _lang('Product Delete  Successfuly'), 'goto' => route('admin.product.index')]);
+                }
+           }
+        }else if($model->payment_option == 'savings'){
+            $invest = Calculation::where('income_id','=',1)->first();
+            $total = $invest->total_income +  $model->oil_total_price;
+            $invest->total_income = $total;
+            $success = $invest->save();
+           if($success){
+               $total_stock = $minus_stock->oil_stack - $model->oil_stack;
+                $minus_stock->oil_stack = $total_stock;
+                $result = $minus_stock->save();
+                if($result){
+                    $models = Product::findOrFail($id);
+                    $models->delete();
+                    return response()->json(['success' => false, 'status' => 'danger', 'message' => _lang('Product Delete  Successfuly'), 'goto' => route('admin.product.index')]);
+                }
+           }
         }
+    }
+
+    public function productreport(){
+        return view('admin.product.product_report');
+    }
+
+    public function stockreport(Request $request){
+        $to_date   = $request->to_date;
+        $form_date = $request->form_date;
+	    $models = Product::where('stack_date','>=', $to_date)->where('stack_date','<=', $form_date)->get();
+        return view('admin.product.stock_list',compact('models'));
     }
